+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Switch to new SQLAlchemy dialect for CrateDB.
 - Removed workaround for `_`-prefixed column names.
   The package now requires CrateDB 6.2 or higher.
+- Dependencies: Updated to vanilla meltanolabs-target-postgres 0.6
 
 ## 2023-12-08 v0.0.1
 - Make it work. It can run the canonical Meltano GitHub -> DB example.